feat: refactor and usefull informative functions
